Implement pure elixir sharded backend.
Utilize pg2 to discover and shard requests to a fleet of genservers,
each which will perform the basic components of the leaky bucket algorithm.

This has two advantages over memcached:

1. No need to spam CAS operations to deal with concurrency safety
2. Can take advantage of erlang native rpcs which are quite efficient.

There is the slight drawback of sequential read/writes in genservers,
but that's frankly equivalent to what is found in a system like redis,
which would have been another backend choice, so eh.